feat(anki): Add tags and status fields to AnkiCard interface
The changes add two new fields, `tags` and `status`, to the `AnkiCard` interface. This allows for better representation of Anki card data, as cards can have associated tags and a status (e.g., new, learning, or reviewed).
